+++ message/name.go
@@ -25,7 +25,9 @@ type NameCollection interface {
 	Each(fn func(Name) bool) bool
 }
 
-// IsIntersectingN returns true if a and b contain any of the same names.
+// IsIntersectingN returns true if a and b are intersecting.
+//
+// That is, it returns true if a and b contain any of the same names.
 func IsIntersectingN(a, b NameCollection) bool {
 	return !a.Each(func(n Name) bool {
 		return !b.Has(n)
@@ -33,6 +35,8 @@ func IsIntersectingN(a, b NameCollection) bool {
 }
 
 // IsSubsetN returns true if sub is a (non-strict) subset of sup.
+//
+// That is, it returns true if sup contains all of the names in sub.
 func IsSubsetN(sub, sup NameCollection) bool {
 	return sub.Each(func(n Name) bool {
 		return sup.Has(n)